+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Detect reconverging grammars in GLR analysis.
+- Separately track shift/reduce history in conflict arcs.
+
 ### Changed
 - Grammar parser is now self-hosted (generated based on a grammar file).
 - Make backend annotations in grammar files optional.
-- Detect reconverging grammars in GLR analysis.
 - Improve grammar rule ID printing.
 
 ## 0.3.0 - 2018-08-10